Add OSD migrations that include disk slot changes
OSD Migration currently only works if the same slot is used. This works
because the generation of proposals is deterministic, so if all disks
are in the same position we end up with the same proposal.

This commit changes that behaviour: instead of creating a new proposal,
the old one is used and adapted. This happens in the following steps:

1) When `salt-run replace.osd X` is used and a proposal file (e.g.
    "data1.ceph.yml" in /srv/pillar/proposals/...) gets the "-replace"
    suffix, a new attribute is written for the disks that get replaced
2) When `salt-run proposals.populate` is used (e.g. in stage.1), all
    proposal files that have the "-replace" suffix get parsed and the
    disks in the proposal are compared to the currently installed disks
    on the node.
3) Disks in the old proposal that have the attribute inserted in 1) and
    are still present (i.e. replaced in place) plus new disks (disks
    that are not in the old proposal but present) are considered free.
4) The disks in the old proposal get interated over and if they are
    labeled for replacement they are swapped with one of the free disks.
5) This modified proposal gets saved in the place of the original one
    and can be used by later stages

Once the replacement in the proposal went through, new proposals are
generated on all other nodes in case replacements and node additions
happen at once.

Signed-off-by: Alexander Graul <agraul@suse.com>
(cherry picked from commit a0196dcec2457915a116eaab5a7b80451029d44a)
Signed-off-by: Alexander Graul <agraul@suse.com>